refactor newsfeed logic
