+++ src/components/ActionBar/style.module.scss
@@ -22,53 +22,3 @@
 .help-row {
 	height: 96px;
 }
-
-.drawer-content {
-	transition-property: all !important;
-	overflow: hidden;
-	overflow: clip;
-}
-
-.drawer-body {
-	height: 100%;
-	display: flex;
-	flex-direction: column;
-	padding: 0;
-}
-
-.reading-header {
-	background-position: center;
-	background-size: cover;
-	background-image: linear-gradient(to bottom, transparent, var(--mantine-color-body)), var(--image-url);
-}
-
-.reading-content {
-	pre {
-		white-space: pre-wrap;
-		background-color: var(--mantine-color-slate-9);
-
-		@include light {
-			background-color: var(--mantine-color-slate-0);
-		}
-	}
-
-	code {
-		background-color: unset !important;
-		color: var(--mantine-color-bright) !important;
-		padding: 0;
-	}
-
-	a {
-		color: var(--mantine-color-surreal-5);
-	}
-}
-
-.back-button {
-	-webkit-backdrop-filter: blur(4px);
-	backdrop-filter: blur(4px);
-	z-index: 10;
-	border-radius: var(--mantine-radius-lg);
-	background-color: rgba(0, 0, 0, 0.35);
-	background-clip: padding-box;
-	border: 1px solid rgba(255, 255, 255, 0.3);
-}
